Allow TBTC contract owner to recover misfunded ERC20/ERC721 tokens
In case someone sends tokens to TBTC token contract address, these
tokens can be recovered by the governance. We observed such cases of
misfunds for tBTC v1 so to become more user-friendly, we are giving
rescue tools to the governance.

In practice, recovery functions will have to be proxied through the
vending machine because the vending machine will be the owner of TBTC
token contract.