Moved approveAndCall implementation to ERC20WithPermit
We want all tokens in this repo - including the ones created in tests -
to have approveAndCall so instead of repeating the same implementation
and tests over and over again, we can put approveAndCall implementation
inside ERC20WithPermit. We already have there burn and burnFrom that are
something beyond ERC20/EIP2612/EIP712.