Do not allow to transfer update initiator to 0x0
If the governance decides to block updates, they may transfer the
initiator role to the contract address. Checking for 0x0 makes sense
given that tooling used for submitting transactions might be imperfect
and surprise governance with some undefined/nulls in the code and
landing with 0x0 in the parameter value.