Add RFC for group selection and key generation
Specify a group selection protocol using a sortition pool,
with the submitter of the expensive sortition transaction being reimbursed,
and the frequency of wallet creation being set by the market
based on the gas costs of sortition and the available reimbursement amounts.

For distributed key generation,
use the protocol from `tss-lib` with a number of attempts over 7 days
to avoid expensive re-sortition even if some operators are inactive.
Misbehaviour accusations are resolved by governance.