Deploy contracts on hardhat network for NPM workflow
We've changed the approach regarding how we prepare tbtc-v2 artifacts
used when developing the the dashboard. The dashboard needs the
contracts to be deployed, so we will deploy them on hardhat.
We're saving the artifacts (`--write true`) in the
`deployments/hardhat`. The `prepare-artifacts` script executed during
the `npm publish ...` step will copy the deployed contracts to the
`/artifacts` folder.`
The addresses of the contracts deployed on hardhat network won't
make sense on the Testnet, but that's not a problem, as they don't
matter.