Bridge.revealDeposit implementation
Full implementation of `Bridge.revealDeposit` function and a draft
implementation of `Bridge.sweep` function with just documentation
and TODOs describing what kind of validation needs to be implemented
to keep the mechanism secure.

Bridge manages BTC deposit and redemption and is increasing and
decreasing balances in the Bank as a result of BTC deposit and
redemption operations.

Depositors send BTC funds to the most-recently-created-wallet of the
bridge using pay-to-script-hash (P2SH) which contains hashed information
about the depositor’s minting Ethereum address. Then, the depositor
reveals their desired Ethereum minting address to the Ethereum chain.
The Bridge listens for these sorts of messages and when it gets one,
it checks the Bitcoin network to make sure the funds line up. If they do,
the off-chain wallet may decide to pick this transaction for sweeping,
and when the sweep operation is confirmed on the Bitcoin network, the
wallet informs the Bridge about the sweep increasing appropriate balances
in the Bank.