TBTC application vault contract
TBTC is a fully Bitcoin-backed ERC-20 token pegged to the price of Bitcoin. It
facilitates Bitcoin holders to act on the Ethereum blockchain and access the
decentralized finance (DeFi) ecosystem. Vault mints and redeems TBTC based on
Bitcoin balances in the Bank. TBTC Vault is the owner of TBTC token contract and
is the only contract minting the token.