IVault interface and its support in Bank and TBTC application
`IVault` is an interface for a smart contract consuming Bank balances allowing
the smart contract to receive Bank balances right after sweeping the deposit by
the Bridge. This method allows the depositor to route their deposit revealed to
the Bridge to the particular smart contract in the same transaction the deposit
is revealed. This way, the depositor does not have to execute additional
transaction after the deposit gets swept by the Bridge.

The base cost of increasing TBTC Application balance in the Bank, without
notifying the application with `onBalanceIncreased` call, and with empty
`depositors` and `depositedAmounts` calldata is 32,986 gas, assuming TBTC
application balance is not 0.

The cost of doing the same for a single depositor in calldata and empty
`onBalanceIncreased` is 39,917 gas. It makes sense given that it's an external
contract call, `onlyBank` modifier needs to be checked and

Assuming some TBTC was minted before, the full cost of `onBalanceIncreased` with
minting and 1 depositor is 80,026 and 109,847 with 2 depositors. The cost
increases linearly from this point, about 30k per each new depositor:
139,680 for 3 depositors, 169,525 for 4 depositors, 199,358 for 5 depositors,
and 348,512 for 10 depositors.