Switch to forked `@summa-tx/bitcoin-spv-sol`
The original `@summa-tx/bitcoin-spv-sol` works
only with Solidity 0.5.x so there is a problem
when there is a need to use it from Solidity
0.8.x. The first approach was copying relevant
parts and adjusting them to newer Solidity
but as a long-term approach we decided
to create a keep-network fork and do the
Solidity 0.8 adjustments within it. Here we
got rid of copied library parts and use that
fork instead.