Addressed slither warnings in Vault contract
Two warnings addressed. First, the order of calls in `mint` was
changes to first emit event and then call Bank contract.
Bank contract is trusted so it does not matter anyway but we can
avoid warnings this way instead of disabling detector. Second,
I disabled detector for onBalaneIncreased loop call of _mint.
This is how it needs to work so that we can save on gas and the
caller of the TX needs to ensure it's something that can be
processed given the block gas limit.