Change "previous sweep" to "main UTXO"
Actually, the Bridge tracks the wallet's main
UTXO which can be updated not only by sweeps
but by redemptions as well. Because of that
"previous sweep" nomenclature is too narrow and
we decide to move to "main UTXO". By the way,
we also add `txOutputIndex` to the main UTXO
information as redemptions, unlike sweeps, can
have multiple outputs.