RFC 1: Request the wallet to move funds on redemption timeout/fraud
Redemption timeout should be considered a wallet heartbeat failure
and the wallet should be requested to move funds to another random
wallet. Same situation is in case of a redemption fraud.

We can not trust the wallet who failed once, thus, the necessity
to move the funds early. This also helps to solve problems with
Solidity implementation of sweep proofs and reusing BTC redemption
address after a timeout.