Refactor of redemption request amount management
Refactored the code responsible for redemption
request amount management in order to make it
easier to understand. Instead of calculating
redeemable and minimal amounts upon request,
the current values of treasury fee and tx max
fee are captured and stored within the request.
All calculations are done within the
`submitRedemptionProof` function. Apart from that,
this change adds the support of the treasury and
packs the outcome of `processRedemptionTxOutputs`
into a structure to avoid "stack too deep" errors.