Redeemer must pass the full script not only hash
If we use the output script payload hash, there
is an ambiguity for 20-byte hashes because
a wallet can satisfy the request using
either P2PKH, P2WPKH and P2SH. That means it
can send funds to three different addresses
and still make the request happy. This is not
something the redeemer want. Because of that,
we require to pass the full script which
encodes the information about type and
corresponds to exactly one address.