RFC 1: Redesigning fraud proofs
Given that a Bitcoin transaction could be so large that proving it on
Ethereum would be impossible, fraud proofs need to be processed with
a challenge-response approach. 

When a wallet unlocks a UTXO it needs to calculate a sighash and
provide a signature over that sighash, one for each unlocked UTXO.
Fraud is reported for a UTXO by providing a sighash along with the
wallet's signature over that sighash. From that moment, the wallet
has a certain time to defend itself against the challenge and prove
that the UTXO was spent in an honest way.

UTXO unlocked by the wallet is spent in a fraudulent way if:
- that unlocked UTXO is a revealed deposit that was not proved as
  swept and can not be proved as swept, or
- that unlocked UTXO was not and is not the main wallet's UTXO.

The wallet is allowed to execute only transactions that are accepted by
the submit sweep proof, submit redemption proof, or submit funds migrated
functions. All other transactions are considered fraud. In other words, if
the wallet signed some UTXO, it needs to use that UTXO, and prove to the
Bridge this is a valid usage. If none of it happens within the timeout the
wallet has to defeat against the challenge, this is considered a fraud.