Compute treasury fee as request percentage
So far, the treasury fee was taken as fixed
satoshi amount from each deposit, regardless
its requested amount. That approach incentivizes
big redemptions and does not fit the profitability
model. This change updates that behavior and
computes the treasury fee as a percentage
of each redemption request. The inital value
of this parameter is `0.05%` just like in tBTV v1.