Integrade Bridge with ECDSA Wallet Registry
Bridge creates wallets with ECDSA Wallet Registry. We implement a stub
function to invoke wallet creation. The process is asynchronous and the
Bridge has to wait untill the wallet is created by the Wallet Registry.
Once it's ready the Wallet Registry is exepcted to call the Bridge to
notify it that the ECDSA wallet is ready.

Bridge uses Bitcoin specific identifier for the wallet's based on
HASH160. Since the Wallet Registry is more generic and is using
keccak256 for wallets identification we need to map those two.