Deposit treasury fee
This change introduces a deposit treasury
fee which works in the same way as the existing
redemption fee. This is achieved using the
governable `depositTreasuryFeeDivisor` parameter
that determines how much of the deposited amount
should be taken as treasury fee. Since that
parameter is governable and may change over time,
the actual fee value is computed and stored
upon deposit reveal. During the sweep proof
submission, the fee is cut from each depositor's
balance and the total collected fee is used to
increase the treasury balance accordingly.
This changeset also involves some refactoring
intended to reduce the number of declared
variables in order to avoid `stack too deep`
error fired by Solidity compiler.