Extract proof validation logic to `BitcoinTx`
Proof validation logic is a first candidate to
be moved out of the `Bridge` contract in order
to reduce its size. It can be placed in the
`BitcoinTx` library which already contains
Bitcoin-specific types so placing the logic
operating on those types seems to be a right
move. Since `BitcoinTx` becomes a linked lib
due to that change, the gas cost of main functions
will change as well as the `Bridge` contract size:

- `revealDeposit`: 87606 -> 87746
- `submitSweepProof`: 249183 -> 254624
- `requestRedemption`: 105723 -> 105774
- `submitRedemptionProof`: 138182 -> 143300
- `Bridge` size: 21.488 KB -> 18.924 KB