Introduce `Wallets` library
This change introduces `Wallets` library that
will take care of integration with ECDSA wallets.
This library will take over all the wallet-related
code currently living in the `Bridge` as well as
new code responsible for creating and closing
wallets. The main motivation behind using a
library is the need of reducing `Bridge` byte
size and achieve better code organization.