Update TODOs of timeout and fraud notifiers
Working on wallet integration is a good
opportunity to update the TODOs of
`notifyRedemptionTimeout` and
`submitRedemptionFraudProof` that are tightly
coupled with wallets.

Regarding timeouts, the wallet-related updates
are now handled by `notifyWalletActionTimedOut`
function exposed by `Wallets` library.

Regarding frauds, the entire timeouts becomes
obsolete and can be removed in favor of:
https://github.com/keep-network/tbtc-v2/pull/133