Refactor `validateRedemptionTxInput` function
The validation logic of the input vector for all
outbound transactions made by the wallet is the
same. That means both redemption and moving
funds transactions can use the same logic
for the first stage of the validation. We
already have `validateRedemptionTxInput` that
validates the input of redemption transaction.
To avoid code duplication, we can just make it
more generic by introducing the concept of
`outbound transaction` and use it for moving
funds as well.