Adjust `terminateWallet` function structure
The convention in the `Wallets` lib is that it
exposes external functions that allow to notify
about various wallet lifecycle events. Those
functions should perform all necessary validation
and call internal functions that do the actual
state transition. This approach is already used
for wallet closure. Here we employ it for wallet
termination as well.