Trust the closing wallet to pick target wallets
The main idea is that if the wallet were malicious, it would probably
just steal the funds outright. The only real reason for the malicious
wallet to transfer the funds to a wrong-but-valid wallet would be one
that gives the malicous wallet some sort of advantage, like *another*
wallet it controls.

For instance, if the malicous operator set controlled *two* wallets, and
wanted to drain the funds from both of them, it could intentionally make
of them fail a heartbeat, send the funds to the other one, and then
steal the funds from just the other wallet and get slashed once instead
of twice.

In this case, the loss of opportunity of slashed tokens relative to the
loss of bitcoin is so small, and the chance of one operator group being
able to control two wallets simultaneously is so small that I don't
think it's worth complicating the wallet closure protocol for v2.0
(maybe for v2.1 or v2.2). Simplify, simplify, simplify.