Extract deposit reveal and bridge state into libraries
A Refactoring of the Bridge contract is required given the contract
size constraints. This change is the first step on this journey.
Deposit reveal functionality is extracted to a separate, externally
linked library. Bridge state is also extracted to a separate structure
to allow passing it easily between external libraries. The state - at
least for now - remains in the Bridge contract and the library does
not have to be deployed separately from the Bridge contract.

Bridge contract size comparison before (main) and after:
|  Bridge               ·     27.528  ·       27.528  │
|  Bridge               ·     25.187  ·        2.341  │

Gas costs comparison before (main) and after:
|  revealDeposit          ·    91842  ·    94484  ·     92493  │
|  revealDeposit          ·    94028  ·    96659  ·     94690  │

|  submitSweepProof       ·   181754  ·   336087  ·    255233  │
|  submitSweepProof       ·   181743  ·   336100  ·    255234  │

|  submitRedemptionProof  ·   117746  ·   194802  ·    163500  │
|  submitRedemptionProof  ·   117680  ·   194572  ·    163334  │