Part of Bridge.submitSweepProof extracted to a separate function
The part of the logic resolving wallet, validating its state and
validating the main wallet UTXO has been moved to a separate function.
This is needed to avoid "stack too deep" problems when refactoring the
code moving sweeping functionality to a separate library, out of the
Bridge contract (to be done in the next commits).

Gas usage increased a little (~164 units worst case):
|  submitSweepProof       ·   181754  ·   336087  ·    255233  │
|  submitSweepProof       ·   181884  ·   336251  ·    255380  │