Extract Bridge moving funds between wallets logic into a library
A refactoring of the Bridge contract was required given the contract size
constraints. This change is another step on this journey. Moving funds between
wallets functionality is extracted to a separate, externally linked library. No
changes to the logic in the moving funds code were applied. At this point, we
are safe with not exceeding Bridge contract size with any more functionalities
we want to add, so it is more about consistency of the code rather than fighting
for more space in the contract.

Bridge contract size comparison before and after:
 |  Bridge               ·     15.495 |
 |  Bridge               ·      9.849 │

Gas costs comparison before and after:
 |  revealDeposit           ·    96005  ·    98636  ·    96667 │
 |  revealDeposit           ·    95933  ·    98564  ·    96595 │
 |  requestRedemption       ·   105448  ·   123275  ·   113159 │
 |  requestRedemption       ·   105469  ·   123302  ·   113185 │
 |  submitSweepProof        ·   196167  ·   350538  ·   269679 │
 |  submitSweepProof        ·   196167  ·   350538  ·   269679 │
 |  submitRedemptionProof   ·   132050  ·   209693  ·   177533 │
 |  submitRedemptionProof   ·   132038  ·   209681  ·   177521 │
 |  submitMovingFundsProof  ·   133248  ·   153426  ·   146278 │
 |  submitMovingFundsProof  ·   138735  ·   158895  ·   151751 │

The gas cost of `submitMovingFundsProof` increased by approximately 5400 gas
units.