Extract Bridge sweep logic into a separate library
A refactoring of the Bridge contract is required given the contract size
constraints. This change is another step on this journey. Sweep
functionality is extracted to a separate, externally linked library. No
changes to the logic in the sweeping code were applied.

Bridge contract size comparison before and after:
 |  Bridge               ·     24.852  ·       24.852  │
 |  Bridge               ·     20.703  ·        4.149  │

Gas costs comparison before and after:
 |  revealDeposit           ·    94028  ·    96659  ·    94690   │
 |  revealDeposit           ·    96041  ·    98672  ·    96703   │
 |  submitSweepProof        ·   190576  ·   344943  ·   264072   │
 |  submitSweepProof        ·   196384  ·   350755  ·   269896   │
 |  submitRedemptionProof   ·   126354  ·   203246  ·   171551   │
 |  submitRedemptionProof   ·   126413  ·   203305  ·   171607   │