Extract Bridge redemption logic into a separate library
A refactoring of the Bridge contract is required given the contract size
constraints. This change is another step on this journey. Redemption
functionality is extracted to a separate, externally linked library. No
changes to the logic in the redemption code were applied.

Bridge contract size comparison before and after:
 |  Bridge               ·     20.826   │
 |  Bridge               ·     15.495   │

Gas costs comparison before and after:
 | revealDeposit          ·   96041  ·   98672  ·   96703  │
 | revealDeposit          ·   96005  ·   98636  ·   96667  │
 | requestRedemption      ·  100601  ·  118446  ·  108329  │
 | requestRedemption      ·  105448  ·  123275  ·  113159  │
 | submitSweepProof       ·  196251  ·  350622  ·  269763  │
 | submitSweepProof       ·  196167  ·  350538  ·  269679  │
 | submitRedemptionProof  ·  126671  ·  204876  ·  172773  │
 | submitRedemptionProof  ·  132050  ·  209693  ·  177533  │

The gas cost of `requestRedemption` and `submitRedemptionProof` increased by
approximately 4800 gas units.