Introduce the wallet `Closing` state
This change introduces an additional `Closing`
state between the existing `MovingFunds` and
`Closed` states. The purpose of that state is
introducing a closing period that will effectively
disable the wallet but also keep it exposed to
fraud challenges for a certain period of time
and reduce the risk of deposit-related frauds
that can be committed by closeable wallets.