Remove unnecessary TODOs
We decided to remove the TODOs because only a selected set of maintainers will be reimbursed
for calling this function, and those maintainers will be working on round-robin duty.
There is a low risk that someone will try to frontrun the call. At the same time,
doing this change would complicate the code so we think it is not worth it.