Remove `@keep-network/ecdsa` files before etherscan verification
In order for the `etherscan-verify` plugin not to pick up wrong files
for verification we have to remove contracts of the dependent projects
stored in the `node_modules`. We've been doing that for
`@keep-network/tbtc`, but recently a new dependency was added, to
`@keep-network/ecdsa` and we need to remove its files as well.