Expand NPM workflow triggers
We want toi publish the new NPM package tagged with `development` every
time changes affecting contracts get merged to `main`. We already
trigger the workflow when changes are applied in
- "solidity/contracts/**"
- "solidity/package.json"
- "solidity/yarn.lock"
but it also makes sense to trigger the workflow when there are changes
in
- "solidity/deploy/**"
- "solidity/hardhat.config.ts"
as they also can affect the content of the exported package with the
contracts.