Deploy Bridge as upgradable
We dpeloy Bridge as upgradable contract. We use hardhat-helpers plugin's
feature implemented in https://github.com/keep-network/hardhat-helpers/pull/14
The upgradable proxy deployment is made with
[@openzeppelin/hardhat-upgrades](https://github.com/OpenZeppelin/openzeppelin-upgrades/tree/master#readme)
Afther the contract is deployed an artifact is stored for the Bridge
contract containing address of the proxy and ABI of the implementation contract.