Add gaps to reserve storage slots
An upgradeable contract should reserve slots in storage for possible future versions of the contract, so if a new variable is added in the future it doesn't brake contiguous stored data.

We add gaps to structures that are used in the contract's storage.

We omitted gaps in structures that are used as memory or calldata functions' arguments.

We don't add gaps to structures that are stored in mappings. The mapping values are [stored in different slots](https://docs.soliditylang.org/en/v0.8.13/internals/layout_in_storage.html#mappings-and-dynamic-arrays) and don't require gaps as the values are not contigious.

Refs:
https://docs.openzeppelin.com/contracts/4.x/upgradeable#storage_gaps
https://forum.openzeppelin.com/t/storage-gaps-in-libraries-and-structs/28171