Register moved funds merge requests
UTXOs which represent moved funds are supposed to
be merged by target wallets but those UTXOs need
to be registered in the system first. This is
because we want to allow merging only the moved
funds and nothing more. This pull request extends
the `submitMovingFundsProof` function which now
registers the merge requests in the system upon
a successful proof.

This change caused the `stack too deep` error
and some refactoring was done to mitigate it.
Specifically, `MovingFundsTxOutputsProcessingInfo`
struct was introduced and some code was extracted
to the `validateMovingFundsTxOutputType` function.