Improve deposit sweep proof logic resiliency
This changeset addresses two TODOs related to
deposit sweep proof logic improvements.

First, it hardens the validation of the single
sweep output. So far, it was possible to use
the P2SH output with a proper 20-byte wallet PKH.
Now, the output is strictly validated and must
always be P2(W)PKH.

Second, the main UTXO detection in the inputs
processing loop was improved. So far, it was
possible to forge the actual input considered
as the main UTXO by using another outpoint of the
same origin transaction. Now, the output index
is checked as well so strict equality is preserved.