Support vaults during deposit sweep
This change adds the ability to support vaults
routing during deposit sweep. The routing can
be done if all swept deposits point to the same
trusted vault indicated by the `vault` parameter
of the `submitDepositSweepFunction`. In that case,
the `Bank` just increases the vault's balance by
the sum of all swept deposits and the vault takes
the responsibility for individual depositors'
balances. If the `vault` parameter is the zero
address or points to a non-trusted vaults,
no routing occurs and the depositor's balances
are updated in the `Bank` individually.