Workaround WalletRegistry deployment
In https://github.com/keep-network/tbtc-v2/pull/242 we switched dependency to `@keep-network/ecdsa` to the one that deploys an upgradable proxy for `WalletRegistry`. With this change the deployment scripts here stopped working with an error:
```
ERROR processing /home/runner/work/tbtc-v2/tbtc-v2/solidity/node_modules/@keep-network/ecdsa/export/deploy/04_deploy_wallet_registry.js:
Error: The requested contract was not found. Make sure the source code is available for compilation
```
See: https://github.com/keep-network/tbtc-v2/runs/6320923006?check_suite_focus=true

After investigation, it turned out that it is a problem with incompatibility in the plugins we use. When I removed dependency to `hardhat-deploy-ethers` and used `@nomiclabs/hardhat-ethers` instead the deployment script worked fine.

This problem needs further investigation and fixing the deployment script to work correctly with the imported deployment of `WalletRegistry` as a proxy contract.

https://github.com/keep-network/tbtc-v2/issues/267