Donation vault
BTC donations are needed to re-balance the
system in some rare cases when the BTC:tBTC peg
can be breached. This change introduces a new
`DonationVault` that is meant to handle BTC
donations. That vault is plugged-in into the
deposit flow. An external actor willing to
make a BTC donation just makes a regular
deposit targeting the `DonationVault`. Once
that deposit is swept by the target wallet,
the depositor's tBTC balance is not increased
in the Bank and the deposit is routed to
the `DonationVault` which ignores it as well.
In effect, deposited BTC remains swept by the
wallet but no tBTC-related side effects take
place.