Allow donations from existing Bank balance
The `DonationVault` should allow making donations
using existing either unapproved and approved
Bank balances. The former is handled via
an implementation of the `receiveBalanceApproval`
function which is supposed to be part of the
`Bank`'s `approveBalanceAndCall` call. The
latter is supported through the `donate` function
that allows making a donation with an already
approved balance.