Document wallet lifecycle
In this PR we're adding the diagrams documenting the wallet lifecycle
and transitions between wallet states.
The diagrams were created using the TikZ package of LaTeX. Having them
in the text form allows to keep the diagrams version-controlled and
reviewable.
The PNG versions of the diagrams were rendered using imagemagick tool.
The PDF versions can be generated by running `pdflatex filename.tikz.`