Removed TODO about validating the wallet used for redemption
The oldest wallet should be used for redemptions and this will be suggested
by the dApp. It does not feel like something we need to enforce on-chain.
If we observe that users notoriously bypass the recommendation, we can always
update Bridge contract to require it. For now, we are optimistic and try to
save on gas.