Aligned punctuation rules used in Solidity documentation
Each `@param` ends with a dot `.` now. It is a different approach from the one we
used in `@keep-core/ecdsa` and `@keep-core/random-beacon` but it is time to
acknowledge `@param` descriptions in `tbtc-v2` are much longer and they often do
form (multiple) sentences.

List items of `@dev` requirements have `,` at the end of the bullet and `.` at
the end of the last bullet. Just a convention.