Introduce moving funds timeout reset delay
The `resetMovingFundsTimeout` function can be
called multiple times in a row if the live
wallets count is zero. This is dangerous since
that function will be reimbursable as a
malicious maintainer can drain the reimbursement
pool significantly. To protect against that,
we introduce a reset delay that will allow
resetting the timeout just before it is about
to be exceeded.