Note on upgradeability for `Bridge.receiveBalanceApproval`
`Bridge` is an upgradeable contract deployed behind a `TransparentUpgradeableProxy`.
Accepting redemption data as bytes provides great flexibility. The `Bridge` is
just like any other contract with a balance approved in the `Bank` and can be
upgraded to another version without being bound to a particular interface
forever. For example, what if 2 years from now, `Bridge` uses a single wallet
with all network operators as key share holders with a weekly key rotation?
This flexibility comes with the cost though - developers integrating their
vaults and dApps with `Bridge` using `approveBalanceAndCall` need to pay extra
attention to `redemptionData` and adjust the code in case the expected structure
of `redemptionData`  changes.