First seize, then terminate the wallet
Terminate wallet call allows the WalletRegistry to clear up the wallet data.
Once the data is cleared up, slashing is not possible because it is not
possible to evaluate wallet signing group member list on-chain. We need to first
slash and then terminate. Keep in mind slashing a terminated wallet is not
possible.

This change is not covered with unit tests but will be covered with integration
tests in PR https://github.com/keep-network/tbtc-v2/pull/314.