Don't use unauthenticated git protocol
The `tbtc-v2` package contains a dependency to `@summa-tx/relay-sol@2.0.2`
package, which downloads one of its sub-dependencies via unathenticated
`git://` protocol. That protocol is no longer supported by GitHub. This
means that in certain situations installation of the package or update
of its dependencies using NPM may result in `The unauthenticated git
protocol on port 9418 is no longer supported` error or errors like
`fatal: unable to connect to github.com:`. That what was happening
during execution of `typescript.yml` workflow after dependency to
the `tbtc-v2` package was added to the `package.json`.

As a workaround, we're adding to the workflow a step which configures
Git to use `https://` protocol instead of `git://` when downloading
files.

More info:
https://github.blog/2021-09-01-improving-git-protocol-security-github/.