Update construction of a redemption key
There was a change that modified how the redemption key is build. Instead of
building it as `keccak256(walletPubKeyHash | redeemerOutputScript)` we now build
it as `keccak256(keccak256(redeemerOutputScript) | walletPubKeyHash)`. The
documentation in ts-lib was updated, but we did not update the values and
bulding redemptions key itself in tests. This commit add necessary changes.