Removing BridgeGovernanceTransferred event
Contract size limit forcing us to leave only the essential part of the
code and trim all the nice haves. This event is a nice to have event.
Bridge governance can be read from the chain. Besided, there's another
event on Governable that duplicates this event and informs about the
transferred Bridge Governance.