Cleaned up submitMovingFundsCommitment unit tests
First, we should not authorize third party maintainer as a caller to
MaintainerProxy. The idea is that anyone can call
`submitMovingFundsCommitment` as long as they are a group member of the
committing group.

Second, removed additional nested contexts because we are not testing
negative paths.

Also, took the opportunity and moved activeWalletMainUtxo from the
global level to requestNewWallet unit tests level where this field is
used.