Set `module` property to be `commonJS`
The rules of `tsconfig` are as follows:
- If `target` is not specified, the default is
  `ES3`
- If `module` is not specified, the default is
  `commonJS` if `target` is either `ES3` or
  `ES5`. If the `target` is something else,
  the module is set to `ES6/ES2015`

We recently switched `target` to be `ES2018`
so the `module` is now `ES6/ES2015` which
raises errors during TypeScript compilation.
In order to get the previous behavior, we
need to set the `module` property to `commonJS`
explicitly.

(source: https://www.typescriptlang.org/tsconfig)