Cleanup Bitcoin transaction hash handling
The `typescript` library handles the Bitcoin
transaction hash in the same way as popular
block explorers and wallets. To make it explicit
and avoid future confusion, we extract a
separate `TransactionHash` type that should
provide more context than just `string`.

By the way, we fix the Ethereum Bridge implementation
as it expects the opposite byte order of
transaction hashes.