Add `_disableInitializers` to `Bridge` contract constructor
The initialization function on the Bridge implementation contract was left
uninitialized. Anyone could call that method and set any parameters they wanted.
Because the Bridge contract is designed to be upgradeable, there was a risk that
future upgrades might introduce vulnerabilities (e.g. by adding
delegatecall/sefdestruct capabilities) that could be exploited by an
unauthorized account calling the initialize method on the implementation
contract. Adding `_disableInitializers` to `Bridge` contract constructor fixes
the problem.