Lower optimizer runs for `BridgeGovernance` contract
The `BridgeGovernance` was very close to its
size limit and recent changes around
`redemptionTxMaxTotalFee` pushes the size
beyond the limit. Here we lower the optimizer
runs to make the contract smaller at the cost
of more expensive calls. We can afford that
because the `BridgeGovernance` contract does not
need to be super gas-efficient.