Replacing 'local' tag with 'allowStubs'
We used 'local' tag to denote that we are on hardhat or development
network which is the "local" network and for this we deployed *Stub
cotracts. However, the "local" tag doesn't fit in CI. This change is one
of the changes across the repos to replace "local" tag with
"allowStubs".