Allowing unlimited contract size for hardhat's network
In the random-beacon's project we have the 'allowUnlimitedContractSize' flag
set to true. /random-beacon is a dependency to this project and the
RandomBeacon.sol can't be deployed to hardhat without this flag, it will
throw an error complaining about the gas issues. The reason why we went
over the limit is because we added BeaconAuthorization library in the
random-beacon@2.0.0-dev.43