Deposit refund locktime validation
Here we introduce a requirement that a deposit
must be revealed respectively earlier than
the moment when the deposit become refundable.
The exact length of the ahead period is defined
by a governable parameter. The reasoning behind
that change is the fact that the wallet must
have a reasonable amount of time to sweep a
revealed deposit and do not compete with a
depositor trying to make a refund.