Remove check of `environment` in deployment job
Right now we only support `goerli` network and we expect that the
workflow (when triggered manually) is always dispatched with this
`environment`. Previously we introduced
`github.event.inputs.environment == 'goerli'` condition to not run the
deploy job if workflow gets accidentally run on a different environment.
But even without this condition we don't risk publishing of a package
with some invalid contracts - deploy will fail either due to
unsupported `environment` or due to incorrect account being used.
Actually, returning error instead of cleanly exiting the workflow may be
a better idea in case wrong `environment` is provided - this will alarm
the scheduler that something went wrong with the deployment.