Bridge reimburses the tx cost of submitMovingFundsCommitment
Given that the Bridge requires `submitMovingFundsCommitment` to be called by one
of the wallet members, this function can not be called and reimbursed via
`MaintainerProxy`. Instead, `Bridge` must be responsible for reimbursing the
cost. This is what is implemented in this commit.

The current logic for `submitMovingFundsCommitment` transaction gas cost
reimbursement implemented in `MaintainerProxy` is not going to work and will be
removed in the very next commit.